Implemented filter by field

+++ server/src/common/entity/Quote.hpp
@@ -13,7 +13,7 @@ namespace quote {
             std::string text;
             std::string authorId;
             std::string image;
-            uint32_t creationDate;
+            uint64_t creationDate;
             std::string context;
         public:
             
@@ -22,14 +22,14 @@ namespace quote {
             void setText(std::string text){ this->text = text; }
             void setAuthorId(std::string authorId){ this->authorId = authorId; }
             void setImage(std::string image){ this->image = image; }
-            void setCreationDate(uint32_t creationDate){ this->creationDate = creationDate; }
+            void setCreationDate(uint64_t creationDate){ this->creationDate = creationDate; }
             void setContext(std::string context){ this->context = context; }
             // GETTERS
             std::string getId(){ return this->id; }
             std::string getText(){ return this->text; }
             std::string getAuthorId(){ return this->authorId; }
             std::string getImage(){ return this->image; }
-            uint32_t getCreationDate(){ return this->creationDate; }
+            uint64_t getCreationDate(){ return this->creationDate; }
             std::string getContext(){ return this->context; }
 
             NLOHMANN_DEFINE_TYPE_INTRUSIVE(Quote, id, text, authorId, image, creationDate, context)
